refactors out functions from controller to model
