+++ server/models/index.js
@@ -1,5 +1,7 @@
 const Certifications = require('./certification.model');
+const Persons = require('./person.model');
 
 module.exports = {
-  Certifications: Certifications   
+  Certifications: Certifications,
+  Persons: Persons
 }
